Adds a 'validate_media_type' option, default: true
Many people have complained that the Media Type Spoof Detection does not
work properly for their files. While whitelisting is available for some
files, it's not a general solution. This option leaves the "secure by
default" of the media checker in place, but allows people to turn it off
completely if they so choose.