Don't assume we have Rails.env if we have Rails
As referenced in #1739

Just because the `Rails` constant is defined, it doesn't mean we're
actually in a Rails app. Since there are people who use Paperclip
outside of Rails, and there's no reason we shouldn't be able to run in
those situations. This commit checks for `Rails.env` instead of just
checking for `Rails` and assuming `Rails.env` works.