HasAttachedFile respects global settings
Normally, all of the settings would be set on `Attachment`, and they
were `deep_merge`d fine. However, if you set the global

    config.paperclip_defaults = { :validate_media_type => false }

then the `HasAttachedFile` class wouldn't pick up on it, because it only
uses things that are defined in the class. This change makes it so that
the default options are merged in to the ones that `HasAttachedFile`
uses when it sets up the attachment.

Fixes #1857